refactor: changes required for the pull request implemented
