+++ controllers/users.controller.js
@@ -76,15 +76,8 @@ class UsersController {
 
   async getUser (ctx, next) {
     if (ctx.method !== 'GET') throw new Error('Method not allowed');
-
-    const user = await this.User.findOne({
-      where: {
-        id: ctx.user.id,
-      },
-    });
-
-    if (user) {
-      ctx.body = filterProps(user.dataValues, ['id', 'email', 'first_name', 'last_name']);
+    if (ctx.user) {
+      ctx.body = ctx.user;
       ctx.status = 200;
       await next();
     } else {
